Create a recipe page
Added a new directory named "recipes" within odin-recipes
(main directory).

Added the mac-and-cheese.html into the recipes folder, with its
respective boilerplate HTML.

In the mac-and-cheese.html added an h1 with the recipe name.

Lastly, in the index.html added a link to the mac-and-cheese.html with
the recipe name within it.