Add new recipe page
Added a new image for the indian butter chicken to the
assets/images folder.

Added a new recipe page for the indian butter chicken
(indian-butter-chicken.html), in the recipes folder.

Added a unordered list in the index.html with the links to the
recipe pages  as items, adding a new one for the indian butter
chicken recipe page.