Add image and remaining recipe sections
Added an image of the recipe dish.

Added a description section with its heading and paragraph.

Added an ingredients section with its necessary heading and lists.

Added a steps section with its necessary heading and list.

Also it has been added a link to the home page, to give the user the
option to return to it.