Add new recipe page
Added a new image for the beef and noodles recipe to the
assets/images folder.

Added a new recipe page for the beef and noodles
(beef-and-noodles.html), in the recipes folder.

Added a new recipe link for the beef and noodles page in the recipes
list located in the index.html.