Add free-to-use images
The images used in this project were not obtain from a free-to-use
images site, so new free-to-use images were added to the images folder
and the previous used were deleted.

Each recipe page was updated with its respective new images but also
a new image was added to the index.html, this because the home page
felt empty without one.

The height for all this new images was defined to 500px.